refactor: extract dashboard and info-page routes into ui-routes module (A2)
Move the dashboard and standalone informational pages verbatim out of
lib/routes-ui.js into lib/ui-routes/dashboard-routes.js:
  GET /admin            (dashboard with stats counters)
  GET /admin/swagger    (API reference)
  GET /admin/legal
  GET /admin/upgrade

getStats and ALLOWED_REDIS_LATENCY were only used by the dashboard route and
are dropped from routes-ui.js imports. Wiring mirrors the other extracted
modules (init({ server, call })).

routes-ui.js: 7813 -> 7643 lines; size ratchet lowered. Route table verified
unchanged; all four GET pages render 200 on a booted server.

Stacked on the batch 3 branch.
